+++ bzt/engine.py
@@ -27,14 +27,12 @@ import tempfile
 import time
 import traceback
 import psutil
-import six
 import yaml
 
 from collections import namedtuple, defaultdict
 from json import encoder
 from yaml.representer import SafeRepresenter
-from six.moves import configparser as ConfigParser
-from six.moves import UserDict as DictMixin
+from bzt.modules.moves import ConfigParser, UserDict as DictMixin, iteritems, string_types, text_type, PY2
 
 from bzt import ManualShutdown, NormalShutdown, get_configs_dir
 from bzt.utils import load_class, to_json, BetterDict, ensure_is_dict, dehumanize_time, is_int
@@ -578,7 +576,7 @@ class Configuration(BetterDict):
         """
         if isinstance(obj, dict):
             result = ''
-            for key, val in six.iteritems(obj):
+            for key, val in iteritems(obj):
                 result += cls.__dict_to_overrides(val, '%s.%s' % (path, key))
             return result
         elif isinstance(obj, list):
@@ -656,7 +654,7 @@ class Configuration(BetterDict):
                 pointer = pointer.get(part)
         self.__ensure_list_capacity(pointer, parts[-1])
         self.log.debug("Applying: %s[%s]=%s", pointer, parts[-1], value)
-        if isinstance(parts[-1], six.string_types) and parts[-1][0] == '^':
+        if isinstance(parts[-1], string_types) and parts[-1][0] == '^':
             del pointer[parts[-1][1:]]
         else:
             if value.isdigit():
@@ -684,8 +682,8 @@ class Configuration(BetterDict):
 
 yaml.add_representer(Configuration, SafeRepresenter.represent_dict)
 yaml.add_representer(BetterDict, SafeRepresenter.represent_dict)
-if six.PY2:
-    yaml.add_representer(six.text_type, SafeRepresenter.represent_unicode)
+if PY2:
+    yaml.add_representer(text_type, SafeRepresenter.represent_unicode)
 
 # dirty hack from http://stackoverflow.com/questions/1447287/format-floats-with-standard-json-module
 encoder.FLOAT_REPR = lambda o: format(o, '.3g')
@@ -846,7 +844,7 @@ class ScenarioExecutor(EngineModule):
         """
         if self.__scenario is None:
             scenario = self.execution.get('scenario', {})
-            if isinstance(scenario, six.string_types):
+            if isinstance(scenario, string_types):
                 scenarios = self.engine.config.get("scenarios")
                 if scenario not in scenarios:
                     raise ValueError("Scenario not found in scenarios: %s" % scenario)